final fixes

+++ backend/src/event/event.service.ts
@@ -3,6 +3,7 @@ import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose'; 
 import { Event } from './interfaces/event.interface';
 import { CreateEventDTO } from './dtos/create-event.dto';
+import { UpdateEventDTO } from './dtos/update-event.dto';
 import { HttpException, HttpStatus } from '@nestjs/common';
 
 @Injectable()
@@ -18,8 +19,14 @@ export class EventService {
         .exec();
     }
 
-    async getEvent(id: string): Promise<Event> {
-        return this.eventModel.findById(id).exec();
+    async getEvent(id: string, cc: string): Promise<Event> {
+
+        const event = await this.eventModel.findById(id).exec();
+        if(event.country !== cc) {
+            throw new HttpException('You are not allowed to view this event', HttpStatus.FORBIDDEN);
+        } else {
+            return event;
+        }
     }
 
     async createEvent(createEventDTO: CreateEventDTO, cc: string): Promise<Event> {
@@ -32,12 +39,20 @@ export class EventService {
         return newEvent.save();
     }
 
-    async updateEvent(id: string, createEventDTO: CreateEventDTO): Promise<Event> {
-        const updatedEvent = await this.eventModel.findByIdAndUpdate(id, createEventDTO, { new: true });
+    async updateEvent(id: string, updateEventDTO: UpdateEventDTO, cc: string): Promise<Event> {
+        const existingEvent = await this.eventModel.findById(id).exec();
+        if (existingEvent.country !== cc) {
+            throw new HttpException('You are not allowed to update this event', HttpStatus.FORBIDDEN);
+        }
+        const updatedEvent = await this.eventModel.findByIdAndUpdate(id, updateEventDTO, { new: true });
         return updatedEvent;        
     }
 
-    async deleteEvent(id: string): Promise<Event> {
+    async deleteEvent(id: string, cc: string): Promise<Event> {
+        const existingEvent = await this.eventModel.findById(id).exec();
+        if (existingEvent.country !== cc) {
+            throw new HttpException('You are not allowed to delete this event', HttpStatus.FORBIDDEN);
+        }
         const deletedEvent = await this.eventModel.findByIdAndDelete(id);
         return deletedEvent;
     }
